Fix SQL stuff and table create stuff

+++ backend/cen-v4.sql
@@ -4,7 +4,7 @@ DROP TABLE IF EXISTS TCNReport;
 CREATE TABLE `TCNReport` (
    `reportVK`	varchar(32) DEFAULT "",
    `report`     varchar(4000) DEFAULT "",
-   `reportTS` int,
+   `reportTS` bigint unsigned,
    PRIMARY KEY (`reportVK`,`reportTS`),
    KEY (`reportTS`)
 );
